+++ Front/src/app/models/student.ts
@@ -1,0 +1,12 @@
+export interface Student {
+    id: number;
+    courseId: number;
+    name: string;
+    gender: boolean;
+    genderName: string;
+    birthade: string;
+    zipcode: number;
+    complement: string;
+    password: string;
+    createdAt: string;
+}