+++ Front/src/app/modules/material.module.ts
@@ -1,9 +1,10 @@
 import { MatSnackBarModule } from "@angular/material/snack-bar";
+import { MatTableModule } from "@angular/material/table";
 import { NgModule } from "@angular/core";
 
 @NgModule({
     declarations: [],
-    imports: [MatSnackBarModule],
+    imports: [MatSnackBarModule, MatTableModule],
     providers: [],
     bootstrap: [],
 })